+++ simple-stack-example-mvp/src/main/java/com/zhuinden/simplestackdemoexamplemvp/presentation/paths/second/SecondView.java
@@ -10,6 +10,9 @@ import android.widget.RelativeLayout;
 import com.squareup.coordinators.Coordinators;
 import com.zhuinden.simplestack.Backstack;
 import com.zhuinden.simplestack.Bundleable;
+import com.zhuinden.simplestackdemoexamplemvp.application.CustomApplication;
+
+import javax.inject.Inject;
 
 /**
  * Created by Owner on 2017. 01. 12..
@@ -41,11 +44,12 @@ public class SecondView
 
     private void init(Context context) {
         if(!isInEditMode()) {
-            backstack = Backstack.get(context);
+            CustomApplication.get().getComponent().inject(this);
             secondKey = Backstack.getKey(context);
         }
     }
 
+    @Inject
     Backstack backstack;
 
     SecondKey secondKey;